requirements.txt
aiogram==2.25.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
python-dotenv==0.21.1
pytz==2022.7.1
yarl==1.8.2
motor==3.1.2
requests==2.27.1